Entrega3/storage/requirements.txt
python-dotenv>=1.0.0 
pymongo>=4.0
numpy